requirements.txt
discord.py
mcstatus
google-api-python-client
